pievienoju data tipus Mortgagecalculator failā, kaut kādus vēl sīkumus pamainiju, pievienoju vēl vienu variable readinput

+++ Main.java
@@ -3,7 +3,7 @@ public class Main {
         while (true) {
             System.out.print("|===> Hipotekārā kredīta kalkulators <===|\n");
 
-            int totalCreditSum = Console.readInput(
+            int totalCreditSum = (int)Console.readInput(
                 "Ievadiet kredīta summu: ", 1_000, 1_000_000, "Kredīta summai jābūt robežās no 1'000 līdz 1'000'000\n"
             );
 
@@ -11,7 +11,7 @@ public class Main {
                 "Ievadiet gada procentu likmi: ", 0, 30, "Procentu likmei jābūt robežās no 0 līdz 30\n"
             );
 
-            int years = Console.readInput(
+            int years = (int)Console.readInput(
                 "Ievadiet periodu gados: ", 1, 30, "Periodam jābūt robežās no 1 līdz 30\n"
             );
